change TurnOnMDMMessage component to generic TurnOnMessage and use in end user auth page
