--- frontend/components/TurnOnMdmMessage/_styles.scss
@@ -1,3 +1,0 @@
-.turn-on-mdm-message {
-  max-width: 520px;
-}
