+++ frontend/components/TurnOnMessage/_styles.scss
@@ -1,0 +1,3 @@
+.turn-on-message {
+  max-width: 520px;
+}
