--- frontend/pages/ManageControlsPage/SetupExperience/cards/EndUserAuthentication/components/RequireEndUserAuth/_styles.scss
@@ -1,11 +1,0 @@
-.require-end-user-auth {
-  
-  .empty-table__container {
-    @include tab-empty-state;
-  }
-
-
-  &__required-message {
-    margin-top: 0;
-  }
-}
