--- frontend/pages/ManageControlsPage/SetupExperience/cards/EndUserAuthentication/components/RequireEndUserAuth/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./RequireEndUserAuth";
