+++ frontend/pages/ManageControlsPage/SetupExperience/cards/InstallSoftware/InstallSoftware.tsx
@@ -27,7 +27,7 @@ import DataError from "components/DataError";
 import Spinner from "components/Spinner";
 import TabNav from "components/TabNav";
 import TabText from "components/TabText";
-import TurnOnMdmMessage from "components/TurnOnMdmMessage";
+import TurnOnMessage from "components/TurnOnMessage";
 import CustomLink from "components/CustomLink";
 
 import AddInstallSoftware from "./components/AddInstallSoftware";
@@ -169,7 +169,7 @@ const InstallSoftware = ({
 
       if (turnOnMdm) {
         return (
-          <TurnOnMdmMessage
+          <TurnOnMessage
             header="Additional configuration required"
             info="To customize, first turn on automatic enrollment."
             buttonText="Turn on"
