+++ frontend/pages/admin/IntegrationsPage/cards/MdmSettings/VppPage/VppPage.tsx
@@ -16,7 +16,7 @@ import Button from "components/buttons/Button";
 import DataError from "components/DataError";
 import Spinner from "components/Spinner";
 import PremiumFeatureMessage from "components/PremiumFeatureMessage";
-import TurnOnMdmMessage from "components/TurnOnMdmMessage";
+import TurnOnMessage from "components/TurnOnMessage";
 import PageDescription from "components/PageDescription";
 import EmptyTable from "components/EmptyTable";
 
@@ -139,7 +139,7 @@ const VppPage = ({ router }: IVppPageProps) => {
 
     if (!config?.mdm.enabled_and_configured) {
       return (
-        <TurnOnMdmMessage
+        <TurnOnMessage
           router={router}
           header="Turn on Apple MDM"
           info=" To install Apple App Store apps purchased through Apple Business
